made sequential states of the device. restarts every reboot

+++ include/IR_Receiver.hpp
@@ -11,7 +11,7 @@ class IRReceiver{
 
         void turnOnTCL();
         void turnOffTCL();
-        void setTCLTemp(int temp);
+        void setTCLTemp(uint8_t temp);
         void turnOnMidea();
         void turnOffMidea();
     private:
